Remove the 'pc' attribute to reduce memory usage
Deleted the 'pc' attribute from the table entries. This roughly halves
the size of the table in memory, making it possible to use a larger
table.

However, the prefetcher can no longer verify that existing table entries
belong to a load instruction. This problem is solved by fetching anyway,
like in a branch-prediction buffer. It is possible that the increased
bus activity will impact performance.