Fix bug where stride was updated incorrectly
The approach described in the article was not implemented correctly. In
particular, the prefetcher would reset the stride on a transition from
STEADY to INITIAL. The stride should NOT be updated in this case, as it
gives the prefetcher a chance to return to the STEADY state immediately
if the next prediction is correct.